requirements.txt
# Core dependencies
numpy>=1.23.0
pandas>=1.5.0
joblib>=1.2.0

# Qdrant client for vector database
qdrant-client>=1.7.0

# Sentence Transformers for embeddings
sentence-transformers>=2.2.0
torch>=1.9.0  # Dependency for sentence-transformers

# Environment variable management
python-dotenv>=0.21.0

# Google Gemini API
google-generativeai>=0.3.0

# Progress bar
tqdm>=4.65.0

# Plotting and visualization
plotly>=5.10.0

# Machine learning utilities
scikit-learn>=1.2.0
scipy>=1.9.0

# Gradio for UI
gradio>=3.50.0

# LangChain and related packages
langchain>=0.0.300
openai>=0.27.0
tiktoken>=0.5.0
langchain-community>=0.0.20
pdfplumber>=0.9.0